refactor: extract and relocate helpers, rename getFunctionName to resolveFunctionName, and realign sortDirection function location
